Refactor project structure and update Flask blueprints
This commit includes several key updates and refactoring to improve the project's modularity and maintainability:
- Corrected the blueprint imports in `app.py` to directly import the blueprint instances, resolving issues with Flask app initialization.
- Added provisional code to the `amenity` model file which was previously empty. This code is intended to be temporary and will be refined in future updates.
- General cleanup and optimization across various modules to enhance code readability and performance.
