+++ app.py
@@ -1,5 +1,6 @@
 from flask import Flask
-from app.api import user_routes, place_routes
+from app.api.user_routes import user_routes
+from app.api.place_routes import place_routes
 from config import SECRET_KEY
 
 def create_app():
@@ -9,12 +10,11 @@ def create_app():
     my_app.config.from_pyfile('config.py')
 
     # Registering blueprints from the api package
-    my_app.register_blueprint(user_routes.blueprint)
-    my_app.register_blueprint(place_routes.blueprint)
+    my_app.register_blueprint(user_routes) # removed blueprint from user_route
+    my_app.register_blueprint(place_routes) # same as above
     
     return my_app
 
 if __name__ == '__main__':
     my_app = create_app()
     my_app.run(debug=True, host='0.0.0.0')
-
