+++ app/api/place_routes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify, request
-from app.models import Place
+from app.models.place import Place
 
 place_routes = Blueprint('place_routes', __name__)
