+++ app/models/__init__.py
@@ -1,6 +1,2 @@
-from .base_model import BaseModel
-from .user import User
-from .place import Place
-from .review import Review
-from .city import City
+
